+++ geometry_wars/ui/GOButton.h
@@ -23,7 +23,7 @@ class GOButton : public GameObject {
     ~GOButton() = default;
     GOButton(float left, float top, float width, float height, std::string label, Color color, std::function<void()> on_click);
     void draw(GameBuffer buffer, Camera* camera);
-    bool is_inside(int x, int y, int buffer_width, int buffer_height);
+    bool is_inside(float x, float y, int buffer_width, int buffer_height);
     void mouse_outside(bool lmb_pressed, bool rmb_pressed);
     void mouse_press(int button);
     void mouse_release(int button);
